playwright/requirements.txt
pytest-playwright
python-dotenv
beautifulsoup4
